tests/requirements_web.txt
Flask==2.3.3
neo4j==5.14.1
python-dotenv==1.0.0
